fix code check and hdi bug fix

+++ bluetooth/hdi/ohos/hardware/bt/v1_0/server/implement/ohos_bt_vendor_lib.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021 Huawei Device Co., Ltd.
+ * Copyright (c) 2021-2022 Huawei Device Co., Ltd.
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -16,7 +16,7 @@
 #ifndef BT_VENDOR_LIB_H
 #define BT_VENDOR_LIB_H
 
-#include <stdint.h>
+#include <cstdint>
 
 #ifdef __cplusplus
 extern "C" {
@@ -130,7 +130,7 @@ typedef void (*free_callback)(void* buf);
  *  HCI Command packet. For example, opcode = 0x0c03 for the HCI_RESET command
  *  packet.
  */
-typedef uint8_t (*cmd_xmit_callback)(uint16_t opcode, void* p_buf);
+typedef size_t (*cmd_xmit_callback)(uint16_t opcode, void* p_buf);
 
 typedef struct {
     /**
